+++ tests/tests.rs
@@ -920,6 +920,8 @@ use render::render;
 #[test] fn e_image_041() { assert_eq!(render("e-image-041"), 0); }
 #[test] fn e_image_042() { assert_eq!(render("e-image-042"), 0); }
 #[test] fn e_image_043() { assert_eq!(render("e-image-043"), 0); }
+#[test] fn e_image_045() { assert_eq!(render("e-image-045"), 0); }
+#[test] fn e_image_046() { assert_eq!(render("e-image-046"), 0); }
 #[test] fn e_line_001() { assert_eq!(render("e-line-001"), 0); }
 #[test] fn e_line_002() { assert_eq!(render("e-line-002"), 0); }
 #[test] fn e_line_003() { assert_eq!(render("e-line-003"), 0); }